+++ libs/rxjs/number/src/lib/is-integer.ts
@@ -11,7 +11,7 @@ import { map } from 'rxjs/operators';
  * [Number.isInteger](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Number/isInteger)
  *
  * @remarks
- * If you want to return the numbers from the check use the [[fromIsInteger]] operator instead
+ * If you want to return the numbers from the check use the [[filterIsInteger]] operator instead
  *
  * @example
  * ```ts