+++ libs/rxjs/string/src/lib/filter-ends-with.ts
@@ -6,7 +6,7 @@ import { MonoTypeOperatorFunction, Observable } from 'rxjs';
 import { filter } from 'rxjs/operators';
 
 /**
- * The `fromEndsWith` operator can be used with an {@link https://rxjs-dev.firebaseapp.com/guide/observable|Observable} string
+ * The `filterEndsWith` operator can be used with an {@link https://rxjs-dev.firebaseapp.com/guide/observable|Observable} string
  * value and returns a string value if the string of length ends with the passed character
  *
  * @param character The character to check the string ends with
@@ -20,13 +20,13 @@ import { filter } from 'rxjs/operators';
  * @example
  * ```ts
  * from(['test', 'testing'])
- *  .pipe(fromEndsWith('g'))
+ *  .pipe(filterEndsWith('g'))
  *  .subscribe(....) // ['testing']
  * ```
  *
  * @returns String that passes the equality check of [String.prototype.endsWith](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/String/endsWith)
  * @category RxJS From String Query
  */
-export function fromEndsWith(character: string, length?: number): MonoTypeOperatorFunction<string> {
+export function filterEndsWith(character: string, length?: number): MonoTypeOperatorFunction<string> {
   return (source: Observable<string>) => source.pipe(filter((value) => value.endsWith(character, length)));
 }