+++ libs/rxjs/string/src/lib/from-char-code.spec.ts
@@ -1,20 +1,12 @@
-import { from } from 'rxjs';
-import { reduce, take } from 'rxjs/operators';
 import { fromCharCode } from './from-char-code';
+import { take } from 'rxjs/operators';
 
 describe('fromCharCode', () => {
-  it('should create a value from an array of values', (done) => {
-    from([35, 36, 37, 38])
-      .pipe(
-        reduce((acc, val) => {
-          acc.push(val);
-          return acc;
-        }, []),
-        fromCharCode(),
-        take(1),
-      )
+  it('should create a string from a passed list of number', (done) => {
+    fromCharCode([65, 66, 67, 68])
+      .pipe(take(1))
       .subscribe({
-        next: (value) => expect(value).toBe('#$%&'),
+        next: (value) => expect(value).toBe('ABCD'),
         complete: () => done(),
       });
   });