+++ libs/rxjs/number/CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.0] - 2020-05-10
+
+### Changed
+
+- The current operators that begin with `from*` are being changed to `filter*` as using `from` has
+  a specific context in RxJS with creating Observable values. The following operators are affected: - `fromIsFinite` to `filterIsFinite` - `fromIsInteger` to `filterIsInteger` - `fromIsSafeInteger` to `filterIsSafeInteger`
+  Because of this change this release is a major bump to `2.0.0`
+
 ## [1.0.3] - 2020-05-06
 
 ### Changed
@@ -17,7 +25,7 @@ Initial Release of the rxjs-number library
 
 ### Added
 
-- `fromIsFinite` - Returns a number value from an Observable number if the `Number.isFinite` equality check it true
+- `filterIsFinite` - Returns a number value from an Observable number if the `Number.isFinite` equality check it true
 - `fromIsInteger` - Returns a number value from an Observable number if the `Number.isInteger` equality check it true
 - `fromIsSafeInteger` - Returns a number value from an Observable number if the `Number.isSafeInteger` equality check it true
 - `isFinite` - Returns a boolean value from an Observable number `Number.isFinite` equality check